LayoutTests/fast/frames/seamless/seamless-designMode-expected.txt

Test that being a seamless iframe does not break designMode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.iframe.contentDocument.body.textContent is "PASS"
